# docs/swagger.yaml
...
definitions:
...
  handler.RespBody:
    properties:
      code:
        description: http code
        type: integer
      data:
        description: response data
      msg:
        description: response message
        type: string
      reason:
        description: reason key
        type: string
    type: object
  install.CheckConfigFileResp:
    properties:
      config_file_exist:
        type: boolean
      db_connection_success:
        type: boolean
      db_table_exist:
        type: boolean
    type: object
  install.CheckDatabaseReq:
    properties:
      db_file:
        type: string
      db_host:
        type: string
      db_name:
        type: string
      db_password:
        type: string
      db_type:
        enum:
        - postgres
        - sqlite3
        - mysql
        type: string
      db_username:
        type: string
    required:
    - db_type
    type: object
  install.InitBaseInfoReq:
    properties:
      contact_email:
        maxLength: 500
        type: string
      email:
        maxLength: 500
        type: string
      lang:
        maxLength: 30
        type: string
      login_required:
        type: boolean
      name:
        maxLength: 30
        type: string
      password:
        maxLength: 32
        minLength: 8
        type: string
      site_name:
        maxLength: 30
        type: string
      site_url:
        maxLength: 512
        type: string
    required:
    - contact_email
    - email
    - lang
    - name
    - password
    - site_name
    - site_url
    type: object
...
  schema.NotificationClearRequest:
    properties:
      type:
        description: inbox achievement
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /answer/api/v1/notification/read/state/all:
    put:
      consumes:
      - application/json
      description: ClearUnRead
      parameters:
      - description: NotificationClearRequest
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/schema.NotificationClearRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.RespBody'
      security:
      - ApiKeyAuth: []
      summary: ClearUnRead
      tags:
      - Notification
...
  /installation/base-info:
    post:
      consumes:
      - application/json
      description: init base info
      parameters:
      - description: InitBaseInfoReq
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/install.InitBaseInfoReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.RespBody'
      summary: init base info
      tags:
      - installation
...
  /installation/db/check:
    post:
      consumes:
      - application/json
      description: check database if exist when installation
      parameters:
      - description: CheckDatabaseReq
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/install.CheckDatabaseReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.RespBody'
            - properties:
                data:
                  $ref: '#/definitions/install.CheckConfigFileResp'
              type: object
      summary: check database if exist when installation
      tags:
      - installation
  /installation/init:
    post:
      consumes:
      - application/json
      description: init environment
      parameters:
      - description: CheckDatabaseReq
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/install.CheckDatabaseReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.RespBody'
      summary: init environment
      tags:
      - installation
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"